+++ src/main/java/com/duckblade/osrs/easyteleports/replacers/diary/KaramjaGloves.java
@@ -1,7 +1,8 @@
-package com.duckblade.osrs.easyteleports.replacers;
+package com.duckblade.osrs.easyteleports.replacers.diary;
 
 import com.duckblade.osrs.easyteleports.EasyTeleportsConfig;
 import com.duckblade.osrs.easyteleports.TeleportReplacement;
+import com.duckblade.osrs.easyteleports.replacers.Replacer;
 import com.google.common.collect.ImmutableList;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +16,7 @@ import java.util.List;
 
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__(@Inject))
-public class GhommalsHilt implements Replacer
+public class KaramjaGloves implements Replacer
 {
 
 	private final List<TeleportReplacement> replacements = new ArrayList<>(5);
@@ -26,11 +27,11 @@ public class GhommalsHilt implements Replacer
 	@Override
 	public void onConfigChanged(EasyTeleportsConfig config)
 	{
-		this.enabled = config.enableGhommalsHilt();
+		this.enabled = config.enableKaramjaGloves();
 		replacements.clear();
 
-		replacements.add(new TeleportReplacement("Trollheim", config.replacementGhommalTrollheim()));
-		replacements.add(new TeleportReplacement("Mor Ul Rek", config.replacementGhommalMorUlRek()));
+		replacements.add(new TeleportReplacement("Gem Mine", config.replacementKaramjaGemMine()));
+		replacements.add(new TeleportReplacement("Slayer Master", config.replacementKaramjaSlayerMaster()));
 	}
 
 	@Override
@@ -42,21 +43,15 @@ public class GhommalsHilt implements Replacer
 	@Override
 	public boolean isApplicableToInventory(int itemId)
 	{
-		return itemId == ItemID.CA_OFFHAND_EASY ||
-			itemId == ItemID.CA_OFFHAND_MEDIUM ||
-			itemId == ItemID.CA_OFFHAND_HARD ||
-			itemId == ItemID.CA_OFFHAND_ELITE ||
-			itemId == ItemID.CA_OFFHAND_MASTER ||
-			itemId == ItemID.CA_OFFHAND_GRANDMASTER ||
-			itemId == ItemID.INFERNAL_DEFENDER_GHOMMAL_5 ||
-			itemId == ItemID.INFERNAL_DEFENDER_GHOMMAL_6 ||
-			itemId == ItemID.INFERNAL_DEFENDER_GHOMMAL_5_TROUVER ||
-			itemId == ItemID.INFERNAL_DEFENDER_GHOMMAL_6_TROUVER;
+		return itemId == ItemID.ATJUN_GLOVES_EASY ||
+			itemId == ItemID.ATJUN_GLOVES_MED ||
+			itemId == ItemID.ATJUN_GLOVES_HARD ||
+			itemId == ItemID.ATJUN_GLOVES_ELITE;
 	}
 
 	@Override
 	public EquipmentInventorySlot getEquipmentSlot()
 	{
-		return EquipmentInventorySlot.SHIELD;
+		return EquipmentInventorySlot.GLOVES;
 	}
 }
