+++ src/main/java/com/duckblade/osrs/easyteleports/replacers/diary/RadasBlessing.java
@@ -1,10 +1,12 @@
-package com.duckblade.osrs.easyteleports.replacers;
+package com.duckblade.osrs.easyteleports.replacers.diary;
 
 import com.duckblade.osrs.easyteleports.EasyTeleportsConfig;
 import com.duckblade.osrs.easyteleports.TeleportReplacement;
+import com.duckblade.osrs.easyteleports.replacers.Replacer;
 import com.google.common.collect.ImmutableList;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import net.runelite.api.EquipmentInventorySlot;
 import net.runelite.api.gameval.ItemID;
 
 import javax.inject.Inject;
@@ -14,7 +16,7 @@ import java.util.List;
 
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__(@Inject))
-public class GrandSeedPod implements Replacer
+public class RadasBlessing implements Replacer
 {
 
 	private final List<TeleportReplacement> replacements = new ArrayList<>(5);
@@ -25,11 +27,11 @@ public class GrandSeedPod implements Replacer
 	@Override
 	public void onConfigChanged(EasyTeleportsConfig config)
 	{
-		this.enabled = config.enableGrandSeedPod();
+		this.enabled = config.enableRadasBlessing();
 		replacements.clear();
 
-		replacements.add(new TeleportReplacement("Launch", config.replacementPodLaunch()));
-		replacements.add(new TeleportReplacement("Squash", config.replacementPodSquash()));
+		replacements.add(new TeleportReplacement("Kourend Woodland", config.replacementRadasKourendWoodland()));
+		replacements.add(new TeleportReplacement("Mount Karuulm", config.replacementRadasMountKaruulm()));
 	}
 
 	@Override
@@ -41,6 +43,15 @@ public class GrandSeedPod implements Replacer
 	@Override
 	public boolean isApplicableToInventory(int itemId)
 	{
-		return itemId == ItemID.ALUFT_SEED_POD;
+		return itemId == ItemID.ZEAH_BLESSING_EASY ||
+			itemId == ItemID.ZEAH_BLESSING_MEDIUM ||
+			itemId == ItemID.ZEAH_BLESSING_HARD ||
+			itemId == ItemID.ZEAH_BLESSING_ELITE;
+	}
+
+	@Override
+	public EquipmentInventorySlot getEquipmentSlot()
+	{
+		return EquipmentInventorySlot.AMMO;
 	}
 }
