+++ src/main/java/com/duckblade/osrs/easyteleports/replacers/jewellery/BurningAmulet.java
@@ -1,7 +1,8 @@
-package com.duckblade.osrs.easyteleports.replacers;
+package com.duckblade.osrs.easyteleports.replacers.jewellery;
 
 import com.duckblade.osrs.easyteleports.EasyTeleportsConfig;
 import com.duckblade.osrs.easyteleports.TeleportReplacement;
+import com.duckblade.osrs.easyteleports.replacers.Replacer;
 import com.google.common.collect.ImmutableList;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -16,10 +17,10 @@ import java.util.List;
 
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__(@Inject))
-public class Camulet implements Replacer
+public class BurningAmulet implements Replacer
 {
 
-	private static final String CAMULET_DIALOGUE_HEADER = "Where would you like to teleport to?";
+	private static final String BURNING_AMULET_DIALOGUE_HEADER = "Teleport to...";
 
 	private final List<TeleportReplacement> replacements = new ArrayList<>(5);
 
@@ -29,18 +30,12 @@ public class Camulet implements Replacer
 	@Override
 	public void onConfigChanged(EasyTeleportsConfig config)
 	{
-		this.enabled = config.enableCamulet();
+		this.enabled = config.enableBurningAmulet();
 
 		replacements.clear();
-		// chat dialog
-		replacements.add(new TeleportReplacement("Enakhra's Temple.", config.replacementCamuletEnakhrasTemple()));
-		replacements.add(new TeleportReplacement("Enakhra's Temple Entrance.", config.replacementCamuletEnakhrasTempleEntrance()));
-		// inventory
-		replacements.add(new TeleportReplacement("Enakhra's Temple", config.replacementCamuletEnakhrasTemple()));
-		replacements.add(new TeleportReplacement("Enakhra's Temple Entrance", config.replacementCamuletEnakhrasTempleEntrance()));
-		// equipped
-		replacements.add(new TeleportReplacement("Temple", config.replacementCamuletEnakhrasTemple()));
-		replacements.add(new TeleportReplacement("Surface", config.replacementCamuletEnakhrasTempleEntrance()));
+		replacements.add(new TeleportReplacement("Chaos Temple", config.replacementBurningChaosTemple()));
+		replacements.add(new TeleportReplacement("Bandit Camp", config.replacementBurningBanditCamp()));
+		replacements.add(new TeleportReplacement("Lava Maze", config.replacementBurningLavaMaze()));
 	}
 
 	@Override
@@ -55,7 +50,7 @@ public class Camulet implements Replacer
 		Widget[] children = root.getChildren();
 		return children != null &&
 			children.length >= 5 &&
-			CAMULET_DIALOGUE_HEADER.equals(children[0].getText());
+			BURNING_AMULET_DIALOGUE_HEADER.equals(children[0].getText());
 	}
 
 	@Override
@@ -67,6 +62,10 @@ public class Camulet implements Replacer
 	@Override
 	public boolean isApplicableToInventory(int itemId)
 	{
-		return itemId == ItemID.CAMULET;
+		return itemId == ItemID.BURNING_AMULET_5 ||
+			itemId == ItemID.BURNING_AMULET_4 ||
+			itemId == ItemID.BURNING_AMULET_3 ||
+			itemId == ItemID.BURNING_AMULET_2 ||
+			itemId == ItemID.BURNING_AMULET_1;
 	}
 }
