+++ src/main/java/com/duckblade/osrs/easyteleports/replacers/other/GhommalsHilt.java
@@ -1,7 +1,8 @@
-package com.duckblade.osrs.easyteleports.replacers;
+package com.duckblade.osrs.easyteleports.replacers.other;
 
 import com.duckblade.osrs.easyteleports.EasyTeleportsConfig;
 import com.duckblade.osrs.easyteleports.TeleportReplacement;
+import com.duckblade.osrs.easyteleports.replacers.Replacer;
 import com.google.common.collect.ImmutableList;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -15,7 +16,7 @@ import java.util.List;
 
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__(@Inject))
-public class RadasBlessing implements Replacer
+public class GhommalsHilt implements Replacer
 {
 
 	private final List<TeleportReplacement> replacements = new ArrayList<>(5);
@@ -26,11 +27,11 @@ public class RadasBlessing implements Replacer
 	@Override
 	public void onConfigChanged(EasyTeleportsConfig config)
 	{
-		this.enabled = config.enableRadasBlessing();
+		this.enabled = config.enableGhommalsHilt();
 		replacements.clear();
 
-		replacements.add(new TeleportReplacement("Kourend Woodland", config.replacementRadasKourendWoodland()));
-		replacements.add(new TeleportReplacement("Mount Karuulm", config.replacementRadasMountKaruulm()));
+		replacements.add(new TeleportReplacement("Trollheim", config.replacementGhommalTrollheim()));
+		replacements.add(new TeleportReplacement("Mor Ul Rek", config.replacementGhommalMorUlRek()));
 	}
 
 	@Override
@@ -42,15 +43,21 @@ public class RadasBlessing implements Replacer
 	@Override
 	public boolean isApplicableToInventory(int itemId)
 	{
-		return itemId == ItemID.ZEAH_BLESSING_EASY ||
-			itemId == ItemID.ZEAH_BLESSING_MEDIUM ||
-			itemId == ItemID.ZEAH_BLESSING_HARD ||
-			itemId == ItemID.ZEAH_BLESSING_ELITE;
+		return itemId == ItemID.CA_OFFHAND_EASY ||
+			itemId == ItemID.CA_OFFHAND_MEDIUM ||
+			itemId == ItemID.CA_OFFHAND_HARD ||
+			itemId == ItemID.CA_OFFHAND_ELITE ||
+			itemId == ItemID.CA_OFFHAND_MASTER ||
+			itemId == ItemID.CA_OFFHAND_GRANDMASTER ||
+			itemId == ItemID.INFERNAL_DEFENDER_GHOMMAL_5 ||
+			itemId == ItemID.INFERNAL_DEFENDER_GHOMMAL_6 ||
+			itemId == ItemID.INFERNAL_DEFENDER_GHOMMAL_5_TROUVER ||
+			itemId == ItemID.INFERNAL_DEFENDER_GHOMMAL_6_TROUVER;
 	}
 
 	@Override
 	public EquipmentInventorySlot getEquipmentSlot()
 	{
-		return EquipmentInventorySlot.AMMO;
+		return EquipmentInventorySlot.SHIELD;
 	}
 }
