+++ src/main/java/com/duckblade/osrs/easyteleports/replacers/other/GrandSeedPod.java
@@ -1,11 +1,11 @@
-package com.duckblade.osrs.easyteleports.replacers;
+package com.duckblade.osrs.easyteleports.replacers.other;
 
 import com.duckblade.osrs.easyteleports.EasyTeleportsConfig;
 import com.duckblade.osrs.easyteleports.TeleportReplacement;
+import com.duckblade.osrs.easyteleports.replacers.Replacer;
 import com.google.common.collect.ImmutableList;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
-import net.runelite.api.EquipmentInventorySlot;
 import net.runelite.api.gameval.ItemID;
 
 import javax.inject.Inject;
@@ -15,7 +15,7 @@ import java.util.List;
 
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__(@Inject))
-public class DesertAmulet implements Replacer
+public class GrandSeedPod implements Replacer
 {
 
 	private final List<TeleportReplacement> replacements = new ArrayList<>(5);
@@ -26,11 +26,11 @@ public class DesertAmulet implements Replacer
 	@Override
 	public void onConfigChanged(EasyTeleportsConfig config)
 	{
-		this.enabled = config.enableDesertAmulet();
-
+		this.enabled = config.enableGrandSeedPod();
 		replacements.clear();
-		replacements.add(new TeleportReplacement("Nardah", config.replacementDesertNardah()));
-		replacements.add(new TeleportReplacement("Kalphite cave", config.replacementDesertKalphiteCave()));
+
+		replacements.add(new TeleportReplacement("Launch", config.replacementPodLaunch()));
+		replacements.add(new TeleportReplacement("Squash", config.replacementPodSquash()));
 	}
 
 	@Override
@@ -40,17 +40,8 @@ public class DesertAmulet implements Replacer
 	}
 
 	@Override
-	public EquipmentInventorySlot getEquipmentSlot()
-	{
-		return EquipmentInventorySlot.AMULET;
-	}
-
-	@Override
 	public boolean isApplicableToInventory(int itemId)
 	{
-		return itemId == ItemID.DESERT_AMULET_EASY ||
-			itemId == ItemID.DESERT_AMULET_MEDIUM ||
-			itemId == ItemID.DESERT_AMULET_HARD ||
-			itemId == ItemID.DESERT_AMULET_ELITE;
+		return itemId == ItemID.ALUFT_SEED_POD;
 	}
 }
