+++ src/main/java/com/duckblade/osrs/easyteleports/replacers/quest/EnchantedLyre.java
@@ -1,7 +1,8 @@
-package com.duckblade.osrs.easyteleports.replacers;
+package com.duckblade.osrs.easyteleports.replacers.quest;
 
 import com.duckblade.osrs.easyteleports.EasyTeleportsConfig;
 import com.duckblade.osrs.easyteleports.TeleportReplacement;
+import com.duckblade.osrs.easyteleports.replacers.Replacer;
 import com.google.common.collect.ImmutableList;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -16,10 +17,10 @@ import java.util.List;
 
 @Singleton
 @RequiredArgsConstructor(onConstructor = @__(@Inject))
-public class GiantsoulAmulet implements Replacer
+public class EnchantedLyre implements Replacer
 {
 
-	private static final String GIANTSOUL_AMULET_DIALOGUE_HEADER = "Where would you like to teleport to?";
+	private static final String ENCHANTED_LYRE_DIALOGUE_HEADER = "Where would you like to teleport to?";
 
 	private final List<TeleportReplacement> replacements = new ArrayList<>(5);
 
@@ -29,18 +30,13 @@ public class GiantsoulAmulet implements Replacer
 	@Override
 	public void onConfigChanged(EasyTeleportsConfig config)
 	{
-		this.enabled = config.enableGiantsoulAmulet();
+		this.enabled = config.enableEnchantedLyre();
 
 		replacements.clear();
-		// chat dialog
-		replacements.add(new TeleportReplacement("Bryophyta's lair", config.replacementBryophyta()));
-		replacements.add(new TeleportReplacement("Obor's lair", config.replacementObor()));
-		replacements.add(new TeleportReplacement("Branda and Eldric's lair", config.replacementBrandaAndEldric()));
-
-		// sub menus
-		replacements.add(new TeleportReplacement("Bryophyta", config.replacementBryophyta()));
-		replacements.add(new TeleportReplacement("Obor", config.replacementObor()));
-		replacements.add(new TeleportReplacement("Branda and Eldric", config.replacementBrandaAndEldric()));
+		replacements.add(new TeleportReplacement("Rellekka", config.replacementLyreRellekka()));
+		replacements.add(new TeleportReplacement("Waterbirth Island", config.replacementLyreWaterbirthIsland()));
+		replacements.add(new TeleportReplacement("Neitiznot", config.replacementLyreNeitiznot()));
+		replacements.add(new TeleportReplacement("Jatizso", config.replacementLyreJatizso()));
 	}
 
 	@Override
@@ -55,18 +51,23 @@ public class GiantsoulAmulet implements Replacer
 		Widget[] children = root.getChildren();
 		return children != null &&
 			children.length >= 5 &&
-			GIANTSOUL_AMULET_DIALOGUE_HEADER.equals(children[0].getText());
+			ENCHANTED_LYRE_DIALOGUE_HEADER.equals(children[0].getText());
 	}
 
 	@Override
 	public EquipmentInventorySlot getEquipmentSlot()
 	{
-		return EquipmentInventorySlot.AMULET;
+		return EquipmentInventorySlot.WEAPON;
 	}
 
 	@Override
 	public boolean isApplicableToInventory(int itemId)
 	{
-		return itemId == ItemID.GIANTSOUL_AMULET_CHARGED;
+		return itemId == ItemID.MAGIC_STRUNG_LYRE ||
+			itemId == ItemID.MAGIC_STRUNG_LYRE_2 ||
+			itemId == ItemID.MAGIC_STRUNG_LYRE_3 ||
+			itemId == ItemID.MAGIC_STRUNG_LYRE_4 ||
+			itemId == ItemID.MAGIC_STRUNG_LYRE_5 ||
+			itemId == ItemID.MAGIC_STRUNG_LYRE_INFINITE;
 	}
 }
